+++ Headers/MSRotateEventHandler.h
@@ -32,7 +32,7 @@
 - (BOOL)calculateIsDraggingInsideLayer:(struct CGPoint)arg1;
 - (BOOL)absoluteMouseMoved:(struct CGPoint)arg1 flags:(unsigned long long)arg2;
 - (BOOL)isDraggingCenterPoint:(struct CGPoint)arg1;
-- (BOOL)absoluteMouseDown:(struct CGPoint)arg1 clickCount:(int)arg2 flags:(unsigned long long)arg3;
+- (BOOL)absoluteMouseDown:(struct CGPoint)arg1 clickCount:(unsigned long long)arg2 flags:(unsigned long long)arg3;
 - (void)handlerWillLoseFocus;
 - (void)handlerGotFocus;
